better error handling; file restructure
