+++ errors/index.js
@@ -1,0 +1,2 @@
+module.exports = require('./KrunkerError');
+module.exports.Errors = require('./Errors');
