+++ package.json
@@ -2,9 +2,9 @@
 	"name": "krunker.js",
 	"version": "1.2.3",
 	"description": "Krunker.io Social Page API",
-	"main": "index.js",
+	"main": "./src/index.js",
 	"scripts": {
-		"test": "npm install && node index.js"
+		"test": "npm install && node src/index.js"
 	},
 	"keywords": [
 		"krunker.js",
